server/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5

psycopg2-binary==2.9.9

python-dotenv==1.0.1

marshmallow==3.21.1
flask-marshmallow==0.15.0

pytest==8.0.2
